+++ oalg-mtx/CHANGELOG.md
@@ -19,6 +19,14 @@ and this project adheres to the
 
 ## [3.2.0.2] - 2025-12-29
 
-### Chainged
+### Chained
 - xStandardOrtOrientation for matrices. Now the dimensions are ranging form 0 to 17.
+
+## [3.2.0.3] - 2025-12-30
+
+### Added
+- instances for XStandardEligibleConeG and XStandardEligibleConeFactorG for matrices over Q.
+
+### Chained
+- xStandardOrtOrientation for matrices. Now the dimensions are ranging form 0 to 13.