+++ oalg-stc/CHANGELOG.md
@@ -13,3 +13,6 @@ and this project adheres to the
 
 ## [3.2.1.0] - 2025-12-25
 - instance for opposite field.
+
+## [3.2.2.0] - 2025-12-25
+- instance for XStandardOrtOrientation Q.